Lev: DataPages: Split off Alt Speed and Distance
Completed datapages
