+++ src/Profiles/Lev/DataPages/ANTPLUS_ProfileLevDataPages.h
@@ -4,6 +4,7 @@
 /* RX */
 #include <Profiles/Lev/DataPages/RX/ANTPLUS_LevSpeedSystemInformation1.h>
 #include <Profiles/Lev/DataPages/RX/ANTPLUS_LevSpeedDistanceInformation.h>
+#include <Profiles/Lev/DataPages/RX/ANTPLUS_LevAltSpeedDistanceInformation.h>
 #include <Profiles/Lev/DataPages/RX/ANTPLUS_LevSpeedSystemInformation2.h>
 #include <Profiles/Lev/DataPages/RX/ANTPLUS_LevBatteryInfo.h>
 #include <Profiles/Lev/DataPages/RX/ANTPLUS_LevCapabilities.h>
